@@ -15,7 +15,7 @@ publish_to: none
 version: 1.2.5+0
 
 environment:
-  sdk: ^2.19.6
+  sdk: ^3.0.0
 
 dependencies:
   flutter:
@@ -40,7 +40,8 @@ dependencies:
   csv:
   string_similarity:
   package_info_plus:
-  bsteeleMusicLib:
+  gamepads:
+  bsteele_music_lib:
     path: ../bsteeleMusicLib/bsteele_music_lib
 
 
