@@ -1,13 +1,13 @@
-import 'package:bsteeleMusicLib/app_logger.dart';
-import 'package:bsteeleMusicLib/songs/chord.dart';
-import 'package:bsteeleMusicLib/songs/chord_component.dart';
-import 'package:bsteeleMusicLib/songs/key.dart' as musical_key;
-import 'package:bsteeleMusicLib/songs/lyric_section.dart';
-import 'package:bsteeleMusicLib/songs/music_constants.dart';
-import 'package:bsteeleMusicLib/songs/pitch.dart';
-import 'package:bsteeleMusicLib/songs/scale_note.dart';
-import 'package:bsteeleMusicLib/songs/song_moment.dart';
-import 'package:bsteeleMusicLib/songs/time_signature.dart';
+import 'package:bsteele_music_lib/app_logger.dart';
+import 'package:bsteele_music_lib/songs/chord.dart';
+import 'package:bsteele_music_lib/songs/chord_component.dart';
+import 'package:bsteele_music_lib/songs/key.dart' as musical_key;
+import 'package:bsteele_music_lib/songs/lyric_section.dart';
+import 'package:bsteele_music_lib/songs/music_constants.dart';
+import 'package:bsteele_music_lib/songs/pitch.dart';
+import 'package:bsteele_music_lib/songs/scale_note.dart';
+import 'package:bsteele_music_lib/songs/song_moment.dart';
+import 'package:bsteele_music_lib/songs/time_signature.dart';
 import 'package:bsteele_music_flutter/app/app.dart';
 import 'package:bsteele_music_flutter/bass_study_tool/sheetMusicPainter.dart';
 import 'package:bsteele_music_flutter/bass_study_tool/sheetNote.dart';
