@@ -1,4 +1,4 @@
-import 'package:bsteeleMusicLib/app_logger.dart';
+import 'package:bsteele_music_lib/app_logger.dart';
 import 'package:bsteele_music_flutter/app/app.dart';
 import 'package:bsteele_music_flutter/app/app_theme.dart';
 import 'package:flutter/material.dart';
