@@ -1,9 +1,9 @@
 import 'dart:async';
 
-import 'package:bsteeleMusicLib/app_logger.dart';
-import 'package:bsteeleMusicLib/songs/song.dart';
-import 'package:bsteeleMusicLib/songs/song_update.dart';
-import 'package:bsteeleMusicLib/util/uri_helper.dart';
+import 'package:bsteele_music_lib/app_logger.dart';
+import 'package:bsteele_music_lib/songs/song.dart';
+import 'package:bsteele_music_lib/songs/song_update.dart';
+import 'package:bsteele_music_lib/util/uri_helper.dart';
 import 'package:bsteele_music_flutter/app/app_theme.dart';
 import 'package:bsteele_music_flutter/screens/player.dart';
 import 'package:flutter/foundation.dart';
