@@ -5,7 +5,7 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:bsteeleMusicLib/app_logger.dart';
+import 'package:bsteele_music_lib/app_logger.dart';
 import 'package:bsteele_music_flutter/main.dart';
 import 'package:bsteele_music_flutter/screens/about.dart';
 import 'package:flutter/material.dart';
@@ -31,7 +31,7 @@ void main() async {
   packageInfo = await PackageInfo.fromPlatform();
 
   testWidgets('about test', (WidgetTester tester) async {
-    tester.binding.window.physicalSizeTestValue = const Size(2 * 1920, 2 * 1080); //  fixme: why so big?
+    //tester.binding.window.physicalSizeTestValue = const Size(2 * 1920, 2 * 1080); //  fixme: why so big?
 
     await tester.runAsync(() async {
       // Build our app and trigger a frame.
