@@ -5,10 +5,11 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:bsteeleMusicLib/app_logger.dart';
-import 'package:bsteeleMusicLib/songs/key.dart' as music_key;
-import 'package:bsteeleMusicLib/songs/music_constants.dart';
-import 'package:bsteeleMusicLib/songs/song.dart';
+import 'package:bsteele_music_lib/app_logger.dart';
+import 'package:bsteele_music_lib/songs/key.dart' as music_key;
+import 'package:bsteele_music_lib/songs/key.dart';
+import 'package:bsteele_music_lib/songs/music_constants.dart';
+import 'package:bsteele_music_lib/songs/song.dart';
 import 'package:bsteele_music_flutter/app/appOptions.dart';
 import 'package:bsteele_music_flutter/app/app_theme.dart';
 import 'package:bsteele_music_flutter/screens/edit.dart';
@@ -48,7 +49,7 @@ void main() async {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets('edit widget test', (WidgetTester tester) async {
-    tester.binding.window.physicalSizeTestValue = const Size(4 * 1920, 4 * 1080);
+    //tester.binding.window.physicalSizeTestValue = const Size(4 * 1920, 4 * 1080);
 
     await tester.pumpWidget(
       MaterialApp(
@@ -131,6 +132,15 @@ void main() async {
       expect(keyDropdownFinder, findsOneWidget);
 
       for (var musicKey in music_key.Key.values) {
+        switch (musicKey.keyEnum) {
+          case KeyEnum.Gb:
+          case KeyEnum.Db:
+          case KeyEnum.Ab:
+            continue;
+          default:
+            break;
+        }
+
         logger.i('   musicKey: ${musicKey.halfStep} ${musicKey.toMarkup()}');
 
         await tester.tap(keyDropdownFinder.first, warnIfMissed: false);
@@ -138,8 +148,13 @@ void main() async {
         await tester.pumpAndSettle();
 
         // debugDumpApp();
+        //
+        // for (final element in find.bySubtype<DropdownMenuItem>().evaluate()) {
+        //   logger.i('element: $element');
+        // }
 
         var key = appKeyCreate(AppKeyEnum.editMusicKey, value: musicKey);
+
         final keySelection = find.byKey(key);
 
         //  two are generated by the stack overlay: expect(keySelection, findsOneWidget);
