@@ -6,12 +6,11 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'dart:collection';
-import 'dart:ui' as ui hide window;
 
-import 'package:bsteeleMusicLib/app_logger.dart';
 import 'package:bsteele_music_flutter/app/app.dart';
 import 'package:bsteele_music_flutter/app/app_theme.dart';
 import 'package:bsteele_music_flutter/main.dart';
+import 'package:bsteele_music_lib/app_logger.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:logger/logger.dart';
@@ -27,9 +26,9 @@ void main() {
   });
 
   testWidgets('Search for song test', (WidgetTester tester) async {
-    tester.binding.window.physicalSizeTestValue = const ui.Size(
-        16 * 1920, //  fixme: why is such a width needed?
-        8 * 1080);
+    // tester.binding.window.physicalSizeTestValue = const ui.Size(
+    //     16 * 1920, //  fixme: why is such a width needed?
+    //     8 * 1080);
 
     // Build our app and trigger a frame.
     await tester.pumpWidget(const BSteeleMusicApp());
